+++ CHANGES.rst
@@ -15,6 +15,10 @@ Release 3.0.9 (2016-4-19)
 
 - Fix no-data in layer for selected point bugs.
 
+- Fix mixing timeseries of nested and parent assets.
+
+- Cancel consecutive calls for timeseries.
+
 
 Release 3.0.8 (2016-4-15)
 ---------------------